feat(cdc): epoch-handoff stamp upgrades the stale-cursor refusal
Phase C4 closes Part C: FileMetadata carries (cdc_epoch, last_seq),
stamped at from_graph (covers plain save on non-durable graphs; disable
records the in-process handoff so a later save carries it; capture-off
saves carry forward — 'epoch 7 ended at 412' stays true). A wrong-epoch
cursor whose epoch matches the stamp is told where it ended and whether
it missed anything (caught-up / quantified-gap / no-claim-after-save
cases, each tested; red-first on the stamp match). Golden digest
unchanged (key omitted when never enabled). io/file.rs split 2487->1842
into the existing io/file/ layout (fast-load sidecar codecs) per the
god-file rule before adding the field.
